integration_layer_interface/requirements_ssot.txt
Flask>=3.0.0
loguru>=0.7.0
pyyaml>=6.0.0
requests>=2.32.3
pika==1.3.2
retry==0.9.2
